+++ NEWS.md
@@ -1,3 +1,13 @@
+# apa 0.3.4.9000
+
+## Bug fixes
+
+* Fix error in `cohen_d` with Hedge's g correction not applying to one-sample
+  t-tests. (@spressi, #15)
+* Add `one_sample` argument to `cohens_d_` to specify if Cohen's d is requested
+  for if providing t and n.
+* Fix a missing escape for percent sign in the documentation of `t_apa`.
+
 # apa 0.3.4
 
 ## Bug fixes